+++ app/src/main/java/org/sopt/dosopttemplate/data/repository/FollowerRepository.kt
@@ -1,0 +1,4 @@
+package org.sopt.dosopttemplate.data.repository
+
+class FollowerRepository {
+}
