+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/HomeActivity.kt
@@ -2,14 +2,13 @@ package org.sopt.dosopttemplate.presentation.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
 import doubleBackPressed
 import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.data.User
 import org.sopt.dosopttemplate.databinding.ActivityHomeBinding
 import org.sopt.dosopttemplate.presentation.main.doandroid.DoAndroidFragment
-import org.sopt.dosopttemplate.presentation.main.followerlist.FollowerListFragment
+import org.sopt.dosopttemplate.presentation.main.follower.FollowerFragment
 import org.sopt.dosopttemplate.presentation.main.home.HomeFragment
 import org.sopt.dosopttemplate.presentation.main.mypage.MyPageFragment
 import org.sopt.dosopttemplate.util.getParcelable
@@ -41,7 +40,7 @@ class HomeActivity : AppCompatActivity() {
                 }
 
                 R.id.menu_follower_list -> {
-                    replaceFragment(FollowerListFragment())
+                    replaceFragment(FollowerFragment())
                     true
                 }
