+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/follower/FollowerViewModel.kt
@@ -1,0 +1,10 @@
+package org.sopt.dosopttemplate.presentation.main.follower
+
+import androidx.lifecycle.ViewModel
+import org.sopt.dosopttemplate.data.repository.FollowerRepository
+
+class FollowerViewModel(val repository: FollowerRepository) : ViewModel(){
+
+
+
+}
