+++ app/src/main/res/layout/fragment_follower.xml
